HW6/sepsis_mdp_master/requirements.txt
cloudpickle==2.2.1
gymnasium==0.27.1
gymnasium-notices==0.0.1
importlib-metadata==6.0.0
jax-jumpy==0.2.0
numpy==1.24.2
PyYAML==6.0
typing_extensions==4.5.0
zipp==3.15.0
